backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.2
django-cors-headers>=4.3
pymysql>=1.1.0
python-dateutil>=2.9.0
channels>=4.2,<5.0
channels-redis>=4.3,<5.0
daphne>=4.1,<5.0
djangorestframework-simplejwt>=5.4,<6.0
python-dotenv
django-redis